Move impl of request handling to dedicated trait impl
