+++ application/comit_node/src/lib.rs
@@ -6,9 +6,11 @@ pub mod bam_ext;
 pub mod btsieve;
 pub mod comit_client;
 pub mod comit_i_routes;
+pub mod comit_server;
 pub mod http_api;
 pub mod libp2p_bam;
 pub mod logging;
+pub mod network;
 pub mod node_id;
 pub mod seed;
 pub mod settings;
